@@ -1,24 +1,26 @@
 module github.com/cloudogu/k8s-ces-setup
 
-go 1.21
+go 1.22.0
 
-toolchain go1.21.1
+toolchain go1.22.4
 
 require (
-	github.com/cloudogu/cesapp-lib v0.12.1
+	github.com/cloudogu/cesapp-lib v0.12.2
 	github.com/cloudogu/k8s-apply-lib v0.4.2
-	github.com/cloudogu/k8s-component-operator v0.7.0
-	github.com/cloudogu/k8s-dogu-operator v0.38.0
+	github.com/cloudogu/k8s-component-operator v1.0.1
+	github.com/cloudogu/k8s-dogu-operator v1.2.0
+	github.com/cloudogu/k8s-registry-lib v0.2.2
 	github.com/gin-gonic/gin v1.9.1
 	github.com/pkg/errors v0.9.1
 	github.com/sirupsen/logrus v1.9.3
-	github.com/stretchr/testify v1.8.4
+	github.com/stretchr/testify v1.9.0
 	github.com/toorop/gin-logrus v0.0.0-20210225092905-2c785434f26f
-	k8s.io/api v0.28.2
-	k8s.io/apimachinery v0.28.2
-	k8s.io/client-go v0.28.2
-	sigs.k8s.io/controller-runtime v0.16.2
-	sigs.k8s.io/yaml v1.3.0
+	k8s.io/api v0.30.0
+	k8s.io/apimachinery v0.30.0
+	k8s.io/client-go v0.30.0
+	oras.land/oras-go v1.2.5
+	sigs.k8s.io/controller-runtime v0.18.2
+	sigs.k8s.io/yaml v1.4.0
 )
 
 replace google.golang.org/grpc => google.golang.org/grpc v1.29.0
