@@ -25,14 +25,11 @@ func TestReadConfig(t *testing.T) {
 		assert.Equal(t, "ecosystem", c.TargetNamespace)
 		assert.Equal(t, "k8s/k8s-component-operator-crd:0.0.2", c.ComponentOperatorCrdChart)
 		assert.Equal(t, "k8s/k8s-component-operator:0.0.2", c.ComponentOperatorChart)
-		assert.Len(t, c.Components, 3)
-		assert.Equal(t, "1.2.3", c.Components["k8s-etcd"].Version)
-		assert.Equal(t, "k8s", c.Components["k8s-etcd"].HelmRepositoryNamespace)
+		assert.Len(t, c.Components, 2)
 		assert.Equal(t, "0.0.1", c.Components["k8s-dogu-operator"].Version)
 		assert.Equal(t, "k8s", c.Components["k8s-dogu-operator"].HelmRepositoryNamespace)
 		assert.Equal(t, "latest", c.Components["k8s-service-discovery"].Version)
 		assert.Equal(t, "k8s", c.Components["k8s-service-discovery"].HelmRepositoryNamespace)
-		assert.Equal(t, "https://etcdc.yaml", c.EtcdClientImageRepo)
 		assert.Equal(t, logrus.DebugLevel, *c.LogLevel)
 	})
 
