@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package core
 
@@ -20,6 +20,10 @@ func (_m *mockFileClient) EXPECT() *mockFileClient_Expecter {
 // Get provides a mock function with given fields: url, username, password
 func (_m *mockFileClient) Get(url string, username string, password string) ([]byte, error) {
 	ret := _m.Called(url, username, password)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Get")
+	}
 
 	var r0 []byte
 	var r1 error
@@ -73,13 +77,12 @@ func (_c *mockFileClient_Get_Call) RunAndReturn(run func(string, string, string)
 	return _c
 }
 
-type mockConstructorTestingTnewMockFileClient interface {
+// newMockFileClient creates a new instance of mockFileClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockFileClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockFileClient creates a new instance of mockFileClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockFileClient(t mockConstructorTestingTnewMockFileClient) *mockFileClient {
+}) *mockFileClient {
 	mock := &mockFileClient{}
 	mock.Mock.Test(t)
 
