@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package patch
 
@@ -25,6 +25,10 @@ func (_m *mockDynClient) EXPECT() *mockDynClient_Expecter {
 // Resource provides a mock function with given fields: resource
 func (_m *mockDynClient) Resource(resource schema.GroupVersionResource) dynamic.NamespaceableResourceInterface {
 	ret := _m.Called(resource)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Resource")
+	}
 
 	var r0 dynamic.NamespaceableResourceInterface
 	if rf, ok := ret.Get(0).(func(schema.GroupVersionResource) dynamic.NamespaceableResourceInterface); ok {
@@ -66,13 +70,12 @@ func (_c *mockDynClient_Resource_Call) RunAndReturn(run func(schema.GroupVersion
 	return _c
 }
 
-type mockConstructorTestingTnewMockDynClient interface {
+// newMockDynClient creates a new instance of mockDynClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockDynClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockDynClient creates a new instance of mockDynClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockDynClient(t mockConstructorTestingTnewMockDynClient) *mockDynClient {
+}) *mockDynClient {
 	mock := &mockDynClient{}
 	mock.Mock.Test(t)
 
