@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package patch
 
@@ -25,6 +25,10 @@ func (_m *mockGvrMapper) EXPECT() *mockGvrMapper_Expecter {
 // KindFor provides a mock function with given fields: resource
 func (_m *mockGvrMapper) KindFor(resource schema.GroupVersionResource) (schema.GroupVersionKind, error) {
 	ret := _m.Called(resource)
+
+	if len(ret) == 0 {
+		panic("no return value specified for KindFor")
+	}
 
 	var r0 schema.GroupVersionKind
 	var r1 error
@@ -77,6 +81,10 @@ func (_c *mockGvrMapper_KindFor_Call) RunAndReturn(run func(schema.GroupVersionR
 // KindsFor provides a mock function with given fields: resource
 func (_m *mockGvrMapper) KindsFor(resource schema.GroupVersionResource) ([]schema.GroupVersionKind, error) {
 	ret := _m.Called(resource)
+
+	if len(ret) == 0 {
+		panic("no return value specified for KindsFor")
+	}
 
 	var r0 []schema.GroupVersionKind
 	var r1 error
@@ -138,6 +146,10 @@ func (_m *mockGvrMapper) RESTMapping(gk schema.GroupKind, versions ...string) (*
 	_ca = append(_ca, gk)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for RESTMapping")
+	}
 
 	var r0 *meta.RESTMapping
 	var r1 error
@@ -208,6 +220,10 @@ func (_m *mockGvrMapper) RESTMappings(gk schema.GroupKind, versions ...string) (
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for RESTMappings")
+	}
+
 	var r0 []*meta.RESTMapping
 	var r1 error
 	if rf, ok := ret.Get(0).(func(schema.GroupKind, ...string) ([]*meta.RESTMapping, error)); ok {
@@ -270,6 +286,10 @@ func (_c *mockGvrMapper_RESTMappings_Call) RunAndReturn(run func(schema.GroupKin
 func (_m *mockGvrMapper) ResourceFor(input schema.GroupVersionResource) (schema.GroupVersionResource, error) {
 	ret := _m.Called(input)
 
+	if len(ret) == 0 {
+		panic("no return value specified for ResourceFor")
+	}
+
 	var r0 schema.GroupVersionResource
 	var r1 error
 	if rf, ok := ret.Get(0).(func(schema.GroupVersionResource) (schema.GroupVersionResource, error)); ok {
@@ -321,6 +341,10 @@ func (_c *mockGvrMapper_ResourceFor_Call) RunAndReturn(run func(schema.GroupVers
 // ResourceSingularizer provides a mock function with given fields: resource
 func (_m *mockGvrMapper) ResourceSingularizer(resource string) (string, error) {
 	ret := _m.Called(resource)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ResourceSingularizer")
+	}
 
 	var r0 string
 	var r1 error
@@ -374,6 +398,10 @@ func (_c *mockGvrMapper_ResourceSingularizer_Call) RunAndReturn(run func(string)
 func (_m *mockGvrMapper) ResourcesFor(input schema.GroupVersionResource) ([]schema.GroupVersionResource, error) {
 	ret := _m.Called(input)
 
+	if len(ret) == 0 {
+		panic("no return value specified for ResourcesFor")
+	}
+
 	var r0 []schema.GroupVersionResource
 	var r1 error
 	if rf, ok := ret.Get(0).(func(schema.GroupVersionResource) ([]schema.GroupVersionResource, error)); ok {
@@ -424,13 +452,12 @@ func (_c *mockGvrMapper_ResourcesFor_Call) RunAndReturn(run func(schema.GroupVer
 	return _c
 }
 
-type mockConstructorTestingTnewMockGvrMapper interface {
+// newMockGvrMapper creates a new instance of mockGvrMapper. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockGvrMapper(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockGvrMapper creates a new instance of mockGvrMapper. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockGvrMapper(t mockConstructorTestingTnewMockGvrMapper) *mockGvrMapper {
+}) *mockGvrMapper {
 	mock := &mockGvrMapper{}
 	mock.Mock.Test(t)
 
