@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package patch
 
@@ -25,6 +25,10 @@ func (_m *mockJsonPatchApplier) EXPECT() *mockJsonPatchApplier_Expecter {
 // Patch provides a mock function with given fields: ctx, jsonPatch, gvk, name
 func (_m *mockJsonPatchApplier) Patch(ctx context.Context, jsonPatch []byte, gvk schema.GroupVersionKind, name string) error {
 	ret := _m.Called(ctx, jsonPatch, gvk, name)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Patch")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, []byte, schema.GroupVersionKind, string) error); ok {
@@ -67,13 +71,12 @@ func (_c *mockJsonPatchApplier_Patch_Call) RunAndReturn(run func(context.Context
 	return _c
 }
 
-type mockConstructorTestingTnewMockJsonPatchApplier interface {
+// newMockJsonPatchApplier creates a new instance of mockJsonPatchApplier. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockJsonPatchApplier(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockJsonPatchApplier creates a new instance of mockJsonPatchApplier. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockJsonPatchApplier(t mockConstructorTestingTnewMockJsonPatchApplier) *mockJsonPatchApplier {
+}) *mockJsonPatchApplier {
 	mock := &mockJsonPatchApplier{}
 	mock.Mock.Test(t)
 
