@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package component
 
@@ -31,6 +31,10 @@ func (_m *mockComponentsClient) EXPECT() *mockComponentsClient_Expecter {
 // AddFinalizer provides a mock function with given fields: ctx, component, finalizer
 func (_m *mockComponentsClient) AddFinalizer(ctx context.Context, component *v1.Component, finalizer string) (*v1.Component, error) {
 	ret := _m.Called(ctx, component, finalizer)
+
+	if len(ret) == 0 {
+		panic("no return value specified for AddFinalizer")
+	}
 
 	var r0 *v1.Component
 	var r1 error
@@ -88,6 +92,10 @@ func (_c *mockComponentsClient_AddFinalizer_Call) RunAndReturn(run func(context.
 func (_m *mockComponentsClient) Create(ctx context.Context, component *v1.Component, opts metav1.CreateOptions) (*v1.Component, error) {
 	ret := _m.Called(ctx, component, opts)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Create")
+	}
+
 	var r0 *v1.Component
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component, metav1.CreateOptions) (*v1.Component, error)); ok {
@@ -144,6 +152,10 @@ func (_c *mockComponentsClient_Create_Call) RunAndReturn(run func(context.Contex
 func (_m *mockComponentsClient) Delete(ctx context.Context, name string, opts metav1.DeleteOptions) error {
 	ret := _m.Called(ctx, name, opts)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Delete")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, string, metav1.DeleteOptions) error); ok {
 		r0 = rf(ctx, name, opts)
@@ -188,6 +200,10 @@ func (_c *mockComponentsClient_Delete_Call) RunAndReturn(run func(context.Contex
 func (_m *mockComponentsClient) DeleteCollection(ctx context.Context, opts metav1.DeleteOptions, listOpts metav1.ListOptions) error {
 	ret := _m.Called(ctx, opts, listOpts)
 
+	if len(ret) == 0 {
+		panic("no return value specified for DeleteCollection")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, metav1.DeleteOptions, metav1.ListOptions) error); ok {
 		r0 = rf(ctx, opts, listOpts)
@@ -231,6 +247,10 @@ func (_c *mockComponentsClient_DeleteCollection_Call) RunAndReturn(run func(cont
 // Get provides a mock function with given fields: ctx, name, opts
 func (_m *mockComponentsClient) Get(ctx context.Context, name string, opts metav1.GetOptions) (*v1.Component, error) {
 	ret := _m.Called(ctx, name, opts)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Get")
+	}
 
 	var r0 *v1.Component
 	var r1 error
@@ -287,6 +307,10 @@ func (_c *mockComponentsClient_Get_Call) RunAndReturn(run func(context.Context, 
 // List provides a mock function with given fields: ctx, opts
 func (_m *mockComponentsClient) List(ctx context.Context, opts metav1.ListOptions) (*v1.ComponentList, error) {
 	ret := _m.Called(ctx, opts)
+
+	if len(ret) == 0 {
+		panic("no return value specified for List")
+	}
 
 	var r0 *v1.ComponentList
 	var r1 error
@@ -349,6 +373,10 @@ func (_m *mockComponentsClient) Patch(ctx context.Context, name string, pt types
 	_ca = append(_ca, ctx, name, pt, data, opts)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Patch")
+	}
 
 	var r0 *v1.Component
 	var r1 error
@@ -416,6 +444,10 @@ func (_c *mockComponentsClient_Patch_Call) RunAndReturn(run func(context.Context
 func (_m *mockComponentsClient) RemoveFinalizer(ctx context.Context, component *v1.Component, finalizer string) (*v1.Component, error) {
 	ret := _m.Called(ctx, component, finalizer)
 
+	if len(ret) == 0 {
+		panic("no return value specified for RemoveFinalizer")
+	}
+
 	var r0 *v1.Component
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component, string) (*v1.Component, error)); ok {
@@ -472,6 +504,10 @@ func (_c *mockComponentsClient_RemoveFinalizer_Call) RunAndReturn(run func(conte
 func (_m *mockComponentsClient) Update(ctx context.Context, component *v1.Component, opts metav1.UpdateOptions) (*v1.Component, error) {
 	ret := _m.Called(ctx, component, opts)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Update")
+	}
+
 	var r0 *v1.Component
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component, metav1.UpdateOptions) (*v1.Component, error)); ok {
@@ -524,9 +560,73 @@ func (_c *mockComponentsClient_Update_Call) RunAndReturn(run func(context.Contex
 	return _c
 }
 
+// UpdateExpectedComponentVersion provides a mock function with given fields: ctx, componentName, version
+func (_m *mockComponentsClient) UpdateExpectedComponentVersion(ctx context.Context, componentName string, version string) (*v1.Component, error) {
+	ret := _m.Called(ctx, componentName, version)
+
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateExpectedComponentVersion")
+	}
+
+	var r0 *v1.Component
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, string, string) (*v1.Component, error)); ok {
+		return rf(ctx, componentName, version)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, string, string) *v1.Component); ok {
+		r0 = rf(ctx, componentName, version)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*v1.Component)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, string, string) error); ok {
+		r1 = rf(ctx, componentName, version)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// mockComponentsClient_UpdateExpectedComponentVersion_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'UpdateExpectedComponentVersion'
+type mockComponentsClient_UpdateExpectedComponentVersion_Call struct {
+	*mock.Call
+}
+
+// UpdateExpectedComponentVersion is a helper method to define mock.On call
+//   - ctx context.Context
+//   - componentName string
+//   - version string
+func (_e *mockComponentsClient_Expecter) UpdateExpectedComponentVersion(ctx interface{}, componentName interface{}, version interface{}) *mockComponentsClient_UpdateExpectedComponentVersion_Call {
+	return &mockComponentsClient_UpdateExpectedComponentVersion_Call{Call: _e.mock.On("UpdateExpectedComponentVersion", ctx, componentName, version)}
+}
+
+func (_c *mockComponentsClient_UpdateExpectedComponentVersion_Call) Run(run func(ctx context.Context, componentName string, version string)) *mockComponentsClient_UpdateExpectedComponentVersion_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		run(args[0].(context.Context), args[1].(string), args[2].(string))
+	})
+	return _c
+}
+
+func (_c *mockComponentsClient_UpdateExpectedComponentVersion_Call) Return(_a0 *v1.Component, _a1 error) *mockComponentsClient_UpdateExpectedComponentVersion_Call {
+	_c.Call.Return(_a0, _a1)
+	return _c
+}
+
+func (_c *mockComponentsClient_UpdateExpectedComponentVersion_Call) RunAndReturn(run func(context.Context, string, string) (*v1.Component, error)) *mockComponentsClient_UpdateExpectedComponentVersion_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
 // UpdateStatus provides a mock function with given fields: ctx, component, opts
 func (_m *mockComponentsClient) UpdateStatus(ctx context.Context, component *v1.Component, opts metav1.UpdateOptions) (*v1.Component, error) {
 	ret := _m.Called(ctx, component, opts)
+
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateStatus")
+	}
 
 	var r0 *v1.Component
 	var r1 error
@@ -584,6 +684,10 @@ func (_c *mockComponentsClient_UpdateStatus_Call) RunAndReturn(run func(context.
 func (_m *mockComponentsClient) UpdateStatusDeleting(ctx context.Context, component *v1.Component) (*v1.Component, error) {
 	ret := _m.Called(ctx, component)
 
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateStatusDeleting")
+	}
+
 	var r0 *v1.Component
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component) (*v1.Component, error)); ok {
@@ -638,6 +742,10 @@ func (_c *mockComponentsClient_UpdateStatusDeleting_Call) RunAndReturn(run func(
 // UpdateStatusInstalled provides a mock function with given fields: ctx, component
 func (_m *mockComponentsClient) UpdateStatusInstalled(ctx context.Context, component *v1.Component) (*v1.Component, error) {
 	ret := _m.Called(ctx, component)
+
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateStatusInstalled")
+	}
 
 	var r0 *v1.Component
 	var r1 error
@@ -694,6 +802,10 @@ func (_c *mockComponentsClient_UpdateStatusInstalled_Call) RunAndReturn(run func
 func (_m *mockComponentsClient) UpdateStatusInstalling(ctx context.Context, component *v1.Component) (*v1.Component, error) {
 	ret := _m.Called(ctx, component)
 
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateStatusInstalling")
+	}
+
 	var r0 *v1.Component
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component) (*v1.Component, error)); ok {
@@ -745,9 +857,72 @@ func (_c *mockComponentsClient_UpdateStatusInstalling_Call) RunAndReturn(run fun
 	return _c
 }
 
+// UpdateStatusNotInstalled provides a mock function with given fields: ctx, component
+func (_m *mockComponentsClient) UpdateStatusNotInstalled(ctx context.Context, component *v1.Component) (*v1.Component, error) {
+	ret := _m.Called(ctx, component)
+
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateStatusNotInstalled")
+	}
+
+	var r0 *v1.Component
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component) (*v1.Component, error)); ok {
+		return rf(ctx, component)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, *v1.Component) *v1.Component); ok {
+		r0 = rf(ctx, component)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*v1.Component)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, *v1.Component) error); ok {
+		r1 = rf(ctx, component)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// mockComponentsClient_UpdateStatusNotInstalled_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'UpdateStatusNotInstalled'
+type mockComponentsClient_UpdateStatusNotInstalled_Call struct {
+	*mock.Call
+}
+
+// UpdateStatusNotInstalled is a helper method to define mock.On call
+//   - ctx context.Context
+//   - component *v1.Component
+func (_e *mockComponentsClient_Expecter) UpdateStatusNotInstalled(ctx interface{}, component interface{}) *mockComponentsClient_UpdateStatusNotInstalled_Call {
+	return &mockComponentsClient_UpdateStatusNotInstalled_Call{Call: _e.mock.On("UpdateStatusNotInstalled", ctx, component)}
+}
+
+func (_c *mockComponentsClient_UpdateStatusNotInstalled_Call) Run(run func(ctx context.Context, component *v1.Component)) *mockComponentsClient_UpdateStatusNotInstalled_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		run(args[0].(context.Context), args[1].(*v1.Component))
+	})
+	return _c
+}
+
+func (_c *mockComponentsClient_UpdateStatusNotInstalled_Call) Return(_a0 *v1.Component, _a1 error) *mockComponentsClient_UpdateStatusNotInstalled_Call {
+	_c.Call.Return(_a0, _a1)
+	return _c
+}
+
+func (_c *mockComponentsClient_UpdateStatusNotInstalled_Call) RunAndReturn(run func(context.Context, *v1.Component) (*v1.Component, error)) *mockComponentsClient_UpdateStatusNotInstalled_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
 // UpdateStatusUpgrading provides a mock function with given fields: ctx, component
 func (_m *mockComponentsClient) UpdateStatusUpgrading(ctx context.Context, component *v1.Component) (*v1.Component, error) {
 	ret := _m.Called(ctx, component)
+
+	if len(ret) == 0 {
+		panic("no return value specified for UpdateStatusUpgrading")
+	}
 
 	var r0 *v1.Component
 	var r1 error
@@ -804,6 +979,10 @@ func (_c *mockComponentsClient_UpdateStatusUpgrading_Call) RunAndReturn(run func
 func (_m *mockComponentsClient) Watch(ctx context.Context, opts metav1.ListOptions) (watch.Interface, error) {
 	ret := _m.Called(ctx, opts)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Watch")
+	}
+
 	var r0 watch.Interface
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, metav1.ListOptions) (watch.Interface, error)); ok {
@@ -855,13 +1034,12 @@ func (_c *mockComponentsClient_Watch_Call) RunAndReturn(run func(context.Context
 	return _c
 }
 
-type mockConstructorTestingTnewMockComponentsClient interface {
+// newMockComponentsClient creates a new instance of mockComponentsClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockComponentsClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockComponentsClient creates a new instance of mockComponentsClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockComponentsClient(t mockConstructorTestingTnewMockComponentsClient) *mockComponentsClient {
+}) *mockComponentsClient {
 	mock := &mockComponentsClient{}
 	mock.Mock.Test(t)
 
