@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package component
 
@@ -26,6 +26,10 @@ func (_m *mockHelmClient) EXPECT() *mockHelmClient_Expecter {
 // InstallOrUpgrade provides a mock function with given fields: ctx, chart
 func (_m *mockHelmClient) InstallOrUpgrade(ctx context.Context, chart *client.ChartSpec) error {
 	ret := _m.Called(ctx, chart)
+
+	if len(ret) == 0 {
+		panic("no return value specified for InstallOrUpgrade")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, *client.ChartSpec) error); ok {
@@ -66,13 +70,12 @@ func (_c *mockHelmClient_InstallOrUpgrade_Call) RunAndReturn(run func(context.Co
 	return _c
 }
 
-type mockConstructorTestingTnewMockHelmClient interface {
+// newMockHelmClient creates a new instance of mockHelmClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockHelmClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockHelmClient creates a new instance of mockHelmClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockHelmClient(t mockConstructorTestingTnewMockHelmClient) *mockHelmClient {
+}) *mockHelmClient {
 	mock := &mockHelmClient{}
 	mock.Mock.Test(t)
 
