@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package data
 
@@ -23,6 +23,10 @@ func (_m *MockRegistryWriter) EXPECT() *MockRegistryWriter_Expecter {
 // WriteConfigToRegistry provides a mock function with given fields: registryConfig
 func (_m *MockRegistryWriter) WriteConfigToRegistry(registryConfig context.CustomKeyValue) error {
 	ret := _m.Called(registryConfig)
+
+	if len(ret) == 0 {
+		panic("no return value specified for WriteConfigToRegistry")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.CustomKeyValue) error); ok {
@@ -62,13 +66,12 @@ func (_c *MockRegistryWriter_WriteConfigToRegistry_Call) RunAndReturn(run func(c
 	return _c
 }
 
-type mockConstructorTestingTNewMockRegistryWriter interface {
+// NewMockRegistryWriter creates a new instance of MockRegistryWriter. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockRegistryWriter(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockRegistryWriter creates a new instance of MockRegistryWriter. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockRegistryWriter(t mockConstructorTestingTNewMockRegistryWriter) *MockRegistryWriter {
+}) *MockRegistryWriter {
 	mock := &MockRegistryWriter{}
 	mock.Mock.Test(t)
 
