@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package data
 
@@ -20,6 +20,10 @@ func (_m *MockSSLGenerator) EXPECT() *MockSSLGenerator_Expecter {
 // GenerateSelfSignedCert provides a mock function with given fields: fqdn, domain, certExpireDays, country, province, locality, altDNSNames
 func (_m *MockSSLGenerator) GenerateSelfSignedCert(fqdn string, domain string, certExpireDays int, country string, province string, locality string, altDNSNames []string) (string, string, error) {
 	ret := _m.Called(fqdn, domain, certExpireDays, country, province, locality, altDNSNames)
+
+	if len(ret) == 0 {
+		panic("no return value specified for GenerateSelfSignedCert")
+	}
 
 	var r0 string
 	var r1 string
@@ -82,13 +86,12 @@ func (_c *MockSSLGenerator_GenerateSelfSignedCert_Call) RunAndReturn(run func(st
 	return _c
 }
 
-type mockConstructorTestingTNewMockSSLGenerator interface {
+// NewMockSSLGenerator creates a new instance of MockSSLGenerator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockSSLGenerator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockSSLGenerator creates a new instance of MockSSLGenerator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockSSLGenerator(t mockConstructorTestingTNewMockSSLGenerator) *MockSSLGenerator {
+}) *MockSSLGenerator {
 	mock := &MockSSLGenerator{}
 	mock.Mock.Test(t)
 
