@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package dogus
 
@@ -20,9 +20,61 @@ func (_m *mockEcoSystemClient) EXPECT() *mockEcoSystemClient_Expecter {
 	return &mockEcoSystemClient_Expecter{mock: &_m.Mock}
 }
 
+// DoguRestarts provides a mock function with given fields: namespace
+func (_m *mockEcoSystemClient) DoguRestarts(namespace string) ecoSystem.DoguRestartInterface {
+	ret := _m.Called(namespace)
+
+	if len(ret) == 0 {
+		panic("no return value specified for DoguRestarts")
+	}
+
+	var r0 ecoSystem.DoguRestartInterface
+	if rf, ok := ret.Get(0).(func(string) ecoSystem.DoguRestartInterface); ok {
+		r0 = rf(namespace)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(ecoSystem.DoguRestartInterface)
+		}
+	}
+
+	return r0
+}
+
+// mockEcoSystemClient_DoguRestarts_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'DoguRestarts'
+type mockEcoSystemClient_DoguRestarts_Call struct {
+	*mock.Call
+}
+
+// DoguRestarts is a helper method to define mock.On call
+//   - namespace string
+func (_e *mockEcoSystemClient_Expecter) DoguRestarts(namespace interface{}) *mockEcoSystemClient_DoguRestarts_Call {
+	return &mockEcoSystemClient_DoguRestarts_Call{Call: _e.mock.On("DoguRestarts", namespace)}
+}
+
+func (_c *mockEcoSystemClient_DoguRestarts_Call) Run(run func(namespace string)) *mockEcoSystemClient_DoguRestarts_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		run(args[0].(string))
+	})
+	return _c
+}
+
+func (_c *mockEcoSystemClient_DoguRestarts_Call) Return(_a0 ecoSystem.DoguRestartInterface) *mockEcoSystemClient_DoguRestarts_Call {
+	_c.Call.Return(_a0)
+	return _c
+}
+
+func (_c *mockEcoSystemClient_DoguRestarts_Call) RunAndReturn(run func(string) ecoSystem.DoguRestartInterface) *mockEcoSystemClient_DoguRestarts_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
 // Dogus provides a mock function with given fields: namespace
 func (_m *mockEcoSystemClient) Dogus(namespace string) ecoSystem.DoguInterface {
 	ret := _m.Called(namespace)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Dogus")
+	}
 
 	var r0 ecoSystem.DoguInterface
 	if rf, ok := ret.Get(0).(func(string) ecoSystem.DoguInterface); ok {
@@ -64,13 +116,12 @@ func (_c *mockEcoSystemClient_Dogus_Call) RunAndReturn(run func(string) ecoSyste
 	return _c
 }
 
-type mockConstructorTestingTnewMockEcoSystemClient interface {
+// newMockEcoSystemClient creates a new instance of mockEcoSystemClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockEcoSystemClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockEcoSystemClient creates a new instance of mockEcoSystemClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockEcoSystemClient(t mockConstructorTestingTnewMockEcoSystemClient) *mockEcoSystemClient {
+}) *mockEcoSystemClient {
 	mock := &mockEcoSystemClient{}
 	mock.Mock.Test(t)
 
