@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package setup
 
@@ -24,6 +24,10 @@ func (_m *MockExecutorStep) EXPECT() *MockExecutorStep_Expecter {
 // GetStepDescription provides a mock function with given fields:
 func (_m *MockExecutorStep) GetStepDescription() string {
 	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetStepDescription")
+	}
 
 	var r0 string
 	if rf, ok := ret.Get(0).(func() string); ok {
@@ -66,6 +70,10 @@ func (_c *MockExecutorStep_GetStepDescription_Call) RunAndReturn(run func() stri
 func (_m *MockExecutorStep) PerformSetupStep(ctx context.Context) error {
 	ret := _m.Called(ctx)
 
+	if len(ret) == 0 {
+		panic("no return value specified for PerformSetupStep")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context) error); ok {
 		r0 = rf(ctx)
@@ -104,13 +112,12 @@ func (_c *MockExecutorStep_PerformSetupStep_Call) RunAndReturn(run func(context.
 	return _c
 }
 
-type mockConstructorTestingTNewMockExecutorStep interface {
+// NewMockExecutorStep creates a new instance of MockExecutorStep. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockExecutorStep(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockExecutorStep creates a new instance of MockExecutorStep. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockExecutorStep(t mockConstructorTestingTNewMockExecutorStep) *MockExecutorStep {
+}) *MockExecutorStep {
 	mock := &MockExecutorStep{}
 	mock.Mock.Test(t)
 
