@@ -1,11 +1,11 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package setup
 
 import (
 	context "context"
 
-	registry "github.com/cloudogu/cesapp-lib/registry"
+	repository "github.com/cloudogu/k8s-registry-lib/repository"
 	mock "github.com/stretchr/testify/mock"
 )
 
@@ -25,6 +25,10 @@ func (_m *MockSetupExecutor) EXPECT() *MockSetupExecutor_Expecter {
 // PerformSetup provides a mock function with given fields: ctx
 func (_m *MockSetupExecutor) PerformSetup(ctx context.Context) (error, string) {
 	ret := _m.Called(ctx)
+
+	if len(ret) == 0 {
+		panic("no return value specified for PerformSetup")
+	}
 
 	var r0 error
 	var r1 string
@@ -78,6 +82,10 @@ func (_c *MockSetupExecutor_PerformSetup_Call) RunAndReturn(run func(context.Con
 func (_m *MockSetupExecutor) RegisterComponentSetupSteps() error {
 	ret := _m.Called()
 
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterComponentSetupSteps")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func() error); ok {
 		r0 = rf()
@@ -115,13 +123,17 @@ func (_c *MockSetupExecutor_RegisterComponentSetupSteps_Call) RunAndReturn(run f
 	return _c
 }
 
-// RegisterDataSetupSteps provides a mock function with given fields: _a0
-func (_m *MockSetupExecutor) RegisterDataSetupSteps(_a0 registry.Registry) error {
-	ret := _m.Called(_a0)
+// RegisterDataSetupSteps provides a mock function with given fields: globalConfig, doguConfigProvider
+func (_m *MockSetupExecutor) RegisterDataSetupSteps(globalConfig *repository.GlobalConfigRepository, doguConfigProvider *repository.DoguConfigRepository) error {
+	ret := _m.Called(globalConfig, doguConfigProvider)
+
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterDataSetupSteps")
+	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(registry.Registry) error); ok {
-		r0 = rf(_a0)
+	if rf, ok := ret.Get(0).(func(*repository.GlobalConfigRepository, *repository.DoguConfigRepository) error); ok {
+		r0 = rf(globalConfig, doguConfigProvider)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -135,14 +147,15 @@ type MockSetupExecutor_RegisterDataSetupSteps_Call struct {
 }
 
 // RegisterDataSetupSteps is a helper method to define mock.On call
-//   - _a0 registry.Registry
-func (_e *MockSetupExecutor_Expecter) RegisterDataSetupSteps(_a0 interface{}) *MockSetupExecutor_RegisterDataSetupSteps_Call {
-	return &MockSetupExecutor_RegisterDataSetupSteps_Call{Call: _e.mock.On("RegisterDataSetupSteps", _a0)}
+//   - globalConfig *repository.GlobalConfigRepository
+//   - doguConfigProvider *repository.DoguConfigRepository
+func (_e *MockSetupExecutor_Expecter) RegisterDataSetupSteps(globalConfig interface{}, doguConfigProvider interface{}) *MockSetupExecutor_RegisterDataSetupSteps_Call {
+	return &MockSetupExecutor_RegisterDataSetupSteps_Call{Call: _e.mock.On("RegisterDataSetupSteps", globalConfig, doguConfigProvider)}
 }
 
-func (_c *MockSetupExecutor_RegisterDataSetupSteps_Call) Run(run func(_a0 registry.Registry)) *MockSetupExecutor_RegisterDataSetupSteps_Call {
+func (_c *MockSetupExecutor_RegisterDataSetupSteps_Call) Run(run func(globalConfig *repository.GlobalConfigRepository, doguConfigProvider *repository.DoguConfigRepository)) *MockSetupExecutor_RegisterDataSetupSteps_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(registry.Registry))
+		run(args[0].(*repository.GlobalConfigRepository), args[1].(*repository.DoguConfigRepository))
 	})
 	return _c
 }
@@ -152,7 +165,7 @@ func (_c *MockSetupExecutor_RegisterDataSetupSteps_Call) Return(_a0 error) *Mock
 	return _c
 }
 
-func (_c *MockSetupExecutor_RegisterDataSetupSteps_Call) RunAndReturn(run func(registry.Registry) error) *MockSetupExecutor_RegisterDataSetupSteps_Call {
+func (_c *MockSetupExecutor_RegisterDataSetupSteps_Call) RunAndReturn(run func(*repository.GlobalConfigRepository, *repository.DoguConfigRepository) error) *MockSetupExecutor_RegisterDataSetupSteps_Call {
 	_c.Call.Return(run)
 	return _c
 }
@@ -160,6 +173,10 @@ func (_c *MockSetupExecutor_RegisterDataSetupSteps_Call) RunAndReturn(run func(r
 // RegisterDoguInstallationSteps provides a mock function with given fields:
 func (_m *MockSetupExecutor) RegisterDoguInstallationSteps() error {
 	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterDoguInstallationSteps")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func() error); ok {
@@ -202,6 +219,10 @@ func (_c *MockSetupExecutor_RegisterDoguInstallationSteps_Call) RunAndReturn(run
 func (_m *MockSetupExecutor) RegisterLoadBalancerFQDNRetrieverSteps() error {
 	ret := _m.Called()
 
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterLoadBalancerFQDNRetrieverSteps")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func() error); ok {
 		r0 = rf()
@@ -242,6 +263,10 @@ func (_c *MockSetupExecutor_RegisterLoadBalancerFQDNRetrieverSteps_Call) RunAndR
 // RegisterSSLGenerationStep provides a mock function with given fields:
 func (_m *MockSetupExecutor) RegisterSSLGenerationStep() error {
 	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterSSLGenerationStep")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func() error); ok {
@@ -284,6 +309,10 @@ func (_c *MockSetupExecutor_RegisterSSLGenerationStep_Call) RunAndReturn(run fun
 func (_m *MockSetupExecutor) RegisterValidationStep() error {
 	ret := _m.Called()
 
+	if len(ret) == 0 {
+		panic("no return value specified for RegisterValidationStep")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func() error); ok {
 		r0 = rf()
@@ -321,13 +350,12 @@ func (_c *MockSetupExecutor_RegisterValidationStep_Call) RunAndReturn(run func()
 	return _c
 }
 
-type mockConstructorTestingTNewMockSetupExecutor interface {
+// NewMockSetupExecutor creates a new instance of MockSetupExecutor. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockSetupExecutor(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockSetupExecutor creates a new instance of MockSetupExecutor. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockSetupExecutor(t mockConstructorTestingTNewMockSetupExecutor) *MockSetupExecutor {
+}) *MockSetupExecutor {
 	mock := &MockSetupExecutor{}
 	mock.Mock.Test(t)
 
