@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package setup
 
@@ -33,6 +33,10 @@ func (_m *mockGinRoutes) Any(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _a0)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Any")
+	}
 
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
@@ -93,6 +97,10 @@ func (_m *mockGinRoutes) DELETE(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes 
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for DELETE")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1...)
@@ -151,6 +159,10 @@ func (_m *mockGinRoutes) GET(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _a0)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for GET")
+	}
 
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
@@ -211,6 +223,10 @@ func (_m *mockGinRoutes) HEAD(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for HEAD")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1...)
@@ -269,6 +285,10 @@ func (_m *mockGinRoutes) Handle(_a0 string, _a1 string, _a2 ...gin.HandlerFunc) 
 	_ca = append(_ca, _a0, _a1)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Handle")
+	}
 
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, string, ...gin.HandlerFunc) gin.IRoutes); ok {
@@ -330,6 +350,10 @@ func (_m *mockGinRoutes) Match(_a0 []string, _a1 string, _a2 ...gin.HandlerFunc)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Match")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func([]string, string, ...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1, _a2...)
@@ -390,6 +414,10 @@ func (_m *mockGinRoutes) OPTIONS(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for OPTIONS")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1...)
@@ -448,6 +476,10 @@ func (_m *mockGinRoutes) PATCH(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _a0)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for PATCH")
+	}
 
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
@@ -508,6 +540,10 @@ func (_m *mockGinRoutes) POST(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for POST")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1...)
@@ -567,6 +603,10 @@ func (_m *mockGinRoutes) PUT(_a0 string, _a1 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for PUT")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, ...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1...)
@@ -619,6 +659,10 @@ func (_c *mockGinRoutes_PUT_Call) RunAndReturn(run func(string, ...gin.HandlerFu
 func (_m *mockGinRoutes) Static(_a0 string, _a1 string) gin.IRoutes {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Static")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, string) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1)
@@ -663,6 +707,10 @@ func (_c *mockGinRoutes_Static_Call) RunAndReturn(run func(string, string) gin.I
 // StaticFS provides a mock function with given fields: _a0, _a1
 func (_m *mockGinRoutes) StaticFS(_a0 string, _a1 http.FileSystem) gin.IRoutes {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for StaticFS")
+	}
 
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, http.FileSystem) gin.IRoutes); ok {
@@ -709,6 +757,10 @@ func (_c *mockGinRoutes_StaticFS_Call) RunAndReturn(run func(string, http.FileSy
 func (_m *mockGinRoutes) StaticFile(_a0 string, _a1 string) gin.IRoutes {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for StaticFile")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, string) gin.IRoutes); ok {
 		r0 = rf(_a0, _a1)
@@ -753,6 +805,10 @@ func (_c *mockGinRoutes_StaticFile_Call) RunAndReturn(run func(string, string) g
 // StaticFileFS provides a mock function with given fields: _a0, _a1, _a2
 func (_m *mockGinRoutes) StaticFileFS(_a0 string, _a1 string, _a2 http.FileSystem) gin.IRoutes {
 	ret := _m.Called(_a0, _a1, _a2)
+
+	if len(ret) == 0 {
+		panic("no return value specified for StaticFileFS")
+	}
 
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(string, string, http.FileSystem) gin.IRoutes); ok {
@@ -806,6 +862,10 @@ func (_m *mockGinRoutes) Use(_a0 ...gin.HandlerFunc) gin.IRoutes {
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Use")
+	}
+
 	var r0 gin.IRoutes
 	if rf, ok := ret.Get(0).(func(...gin.HandlerFunc) gin.IRoutes); ok {
 		r0 = rf(_a0...)
@@ -853,13 +913,12 @@ func (_c *mockGinRoutes_Use_Call) RunAndReturn(run func(...gin.HandlerFunc) gin.
 	return _c
 }
 
-type mockConstructorTestingTnewMockGinRoutes interface {
+// newMockGinRoutes creates a new instance of mockGinRoutes. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockGinRoutes(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockGinRoutes creates a new instance of mockGinRoutes. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockGinRoutes(t mockConstructorTestingTnewMockGinRoutes) *mockGinRoutes {
+}) *mockGinRoutes {
 	mock := &mockGinRoutes{}
 	mock.Mock.Test(t)
 
