@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package setup
 
@@ -23,6 +23,10 @@ func (_m *mockResourcePatchConfigurationValidator) EXPECT() *mockResourcePatchCo
 // Validate provides a mock function with given fields: resourcePatchConfig
 func (_m *mockResourcePatchConfigurationValidator) Validate(resourcePatchConfig []patch.ResourcePatch) error {
 	ret := _m.Called(resourcePatchConfig)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Validate")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func([]patch.ResourcePatch) error); ok {
@@ -62,13 +66,12 @@ func (_c *mockResourcePatchConfigurationValidator_Validate_Call) RunAndReturn(ru
 	return _c
 }
 
-type mockConstructorTestingTnewMockResourcePatchConfigurationValidator interface {
+// newMockResourcePatchConfigurationValidator creates a new instance of mockResourcePatchConfigurationValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockResourcePatchConfigurationValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockResourcePatchConfigurationValidator creates a new instance of mockResourcePatchConfigurationValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockResourcePatchConfigurationValidator(t mockConstructorTestingTnewMockResourcePatchConfigurationValidator) *mockResourcePatchConfigurationValidator {
+}) *mockResourcePatchConfigurationValidator {
 	mock := &mockResourcePatchConfigurationValidator{}
 	mock.Mock.Test(t)
 
