@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package setup
 
@@ -25,6 +25,10 @@ func (_m *mockResourcePatcher) EXPECT() *mockResourcePatcher_Expecter {
 // Patch provides a mock function with given fields: ctx, phase, patches
 func (_m *mockResourcePatcher) Patch(ctx context.Context, phase patch.Phase, patches []patch.ResourcePatch) error {
 	ret := _m.Called(ctx, phase, patches)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Patch")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, patch.Phase, []patch.ResourcePatch) error); ok {
@@ -66,13 +70,12 @@ func (_c *mockResourcePatcher_Patch_Call) RunAndReturn(run func(context.Context,
 	return _c
 }
 
-type mockConstructorTestingTnewMockResourcePatcher interface {
+// newMockResourcePatcher creates a new instance of mockResourcePatcher. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockResourcePatcher(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockResourcePatcher creates a new instance of mockResourcePatcher. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockResourcePatcher(t mockConstructorTestingTnewMockResourcePatcher) *mockResourcePatcher {
+}) *mockResourcePatcher {
 	mock := &mockResourcePatcher{}
 	mock.Mock.Test(t)
 
