@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package setup
 
@@ -23,6 +23,10 @@ func (_m *mockSetupJsonConfigurationValidator) EXPECT() *mockSetupJsonConfigurat
 // Validate provides a mock function with given fields: setupJson
 func (_m *mockSetupJsonConfigurationValidator) Validate(setupJson *context.SetupJsonConfiguration) error {
 	ret := _m.Called(setupJson)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Validate")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(*context.SetupJsonConfiguration) error); ok {
@@ -62,13 +66,12 @@ func (_c *mockSetupJsonConfigurationValidator_Validate_Call) RunAndReturn(run fu
 	return _c
 }
 
-type mockConstructorTestingTnewMockSetupJsonConfigurationValidator interface {
+// newMockSetupJsonConfigurationValidator creates a new instance of mockSetupJsonConfigurationValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func newMockSetupJsonConfigurationValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// newMockSetupJsonConfigurationValidator creates a new instance of mockSetupJsonConfigurationValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func newMockSetupJsonConfigurationValidator(t mockConstructorTestingTnewMockSetupJsonConfigurationValidator) *mockSetupJsonConfigurationValidator {
+}) *mockSetupJsonConfigurationValidator {
 	mock := &mockSetupJsonConfigurationValidator{}
 	mock.Mock.Test(t)
 
