@@ -3,13 +3,11 @@ package setup
 import (
 	"context"
 	"fmt"
-	"github.com/cloudogu/k8s-ces-setup/app/cesregistry"
-
+	k8sreg "github.com/cloudogu/k8s-registry-lib/repository"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 
-	"github.com/cloudogu/cesapp-lib/registry"
 	appcontext "github.com/cloudogu/k8s-ces-setup/app/context"
 )
 
@@ -24,7 +22,7 @@ type SetupExecutor interface {
 	// RegisterComponentSetupSteps adds all setup steps responsible to install vital components into the ecosystem.
 	RegisterComponentSetupSteps() error
 	// RegisterDataSetupSteps adds all setup steps responsible to read, write, or verify data needed by the setup.
-	RegisterDataSetupSteps(registry.Registry) error
+	RegisterDataSetupSteps(globalConfig *k8sreg.GlobalConfigRepository, doguConfigProvider *k8sreg.DoguConfigRepository) error
 	// RegisterDoguInstallationSteps creates install steps for the dogu install list
 	RegisterDoguInstallationSteps() error
 	// PerformSetup starts the setup and executes all registered setup steps
@@ -33,12 +31,13 @@ type SetupExecutor interface {
 
 // Starter is used to init and start the setup process
 type Starter struct {
-	EtcdRegistry  registry.Registry
-	ClientSet     kubernetes.Interface
-	ClusterConfig *rest.Config
-	SetupContext  *appcontext.SetupContext
-	Namespace     string
-	SetupExecutor SetupExecutor
+	globalConfigRepo *k8sreg.GlobalConfigRepository
+	doguConfigRepo    *k8sreg.DoguConfigRepository
+	ClientSet      kubernetes.Interface
+	ClusterConfig  *rest.Config
+	SetupContext   *appcontext.SetupContext
+	Namespace      string
+	SetupExecutor  SetupExecutor
 }
 
 // NewStarter creates a new setup starter struct which one inits registries and starts the setup process
@@ -49,10 +48,9 @@ func NewStarter(ctx context.Context, clusterConfig *rest.Config, k8sClient kuber
 	}
 
 	namespace := setupContext.AppConfig.TargetNamespace
-	etcdRegistry, err := cesregistry.CreateEtcd(namespace)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create registry: %w", err)
-	}
+	cmClient := k8sClient.CoreV1().ConfigMaps(namespace)
+	doguConfigRepo := k8sreg.NewDoguConfigRepository(cmClient)
+	globalConfigRepo := k8sreg.NewGlobalConfigRepository(cmClient)
 
 	setupExecutor, err := NewExecutor(clusterConfig, k8sClient, setupContext)
 	if err != nil {
@@ -60,12 +58,13 @@ func NewStarter(ctx context.Context, clusterConfig *rest.Config, k8sClient kuber
 	}
 
 	return &Starter{
-		EtcdRegistry:  etcdRegistry,
-		ClientSet:     k8sClient,
-		ClusterConfig: clusterConfig,
-		SetupContext:  setupContext,
-		Namespace:     namespace,
-		SetupExecutor: setupExecutor,
+		globalConfigRepo: globalConfigRepo,
+		doguConfigRepo:   doguConfigRepo,
+		ClientSet:        k8sClient,
+		ClusterConfig:    clusterConfig,
+		SetupContext:     setupContext,
+		Namespace:        namespace,
+		SetupExecutor:    setupExecutor,
 	}, nil
 }
 
@@ -76,7 +75,7 @@ func (s *Starter) StartSetup(ctx context.Context) error {
 		return err
 	}
 
-	err = registerSteps(s.SetupExecutor, s.EtcdRegistry, s.SetupContext)
+	err = registerSteps(s.SetupExecutor, s.globalConfigRepo, s.doguConfigRepo, s.SetupContext)
 	if err != nil {
 		return err
 	}
@@ -94,7 +93,7 @@ func (s *Starter) StartSetup(ctx context.Context) error {
 	return nil
 }
 
-func registerSteps(setupExecutor SetupExecutor, etcdRegistry registry.Registry, setupContext *appcontext.SetupContext) error {
+func registerSteps(setupExecutor SetupExecutor, globalConfig *k8sreg.GlobalConfigRepository, doguConfig *k8sreg.DoguConfigRepository, setupContext *appcontext.SetupContext) error {
 	err := setupExecutor.RegisterLoadBalancerFQDNRetrieverSteps()
 	if err != nil {
 		return fmt.Errorf("failed to register steps for creating loadbalancer and retrieving its ip as fqdn: %w", err)
@@ -117,7 +116,7 @@ func registerSteps(setupExecutor SetupExecutor, etcdRegistry registry.Registry, 
 		return fmt.Errorf("failed to register component setup steps: %w", err)
 	}
 
-	err = setupExecutor.RegisterDataSetupSteps(etcdRegistry)
+	err = setupExecutor.RegisterDataSetupSteps(globalConfig, doguConfig)
 	if err != nil {
 		return fmt.Errorf("failed to register data setup steps: %w", err)
 	}
