@@ -27,7 +27,7 @@ func TestStarter_StartSetup(t *testing.T) {
 		expect.RegisterSSLGenerationStep().Return(nil)
 		expect.RegisterValidationStep().Return(nil)
 		expect.RegisterComponentSetupSteps().Return(nil)
-		expect.RegisterDataSetupSteps(mock.Anything).Return(nil)
+		expect.RegisterDataSetupSteps(mock.Anything, mock.Anything).Return(nil)
 		expect.RegisterDoguInstallationSteps().Return(nil)
 		expect.PerformSetup(testCtx).Return(nil, "")
 		starter.SetupExecutor = executorMock
@@ -48,7 +48,7 @@ func TestStarter_StartSetup(t *testing.T) {
 		expect.RegisterSSLGenerationStep().Return(nil)
 		expect.RegisterValidationStep().Return(nil)
 		expect.RegisterComponentSetupSteps().Return(nil)
-		expect.RegisterDataSetupSteps(mock.Anything).Return(nil)
+		expect.RegisterDataSetupSteps(mock.Anything, mock.Anything).Return(nil)
 		expect.RegisterDoguInstallationSteps().Return(nil)
 		expect.PerformSetup(testCtx).Return(nil, "")
 		starter.SetupExecutor = executorMock
@@ -168,7 +168,7 @@ func TestStarter_StartSetup(t *testing.T) {
 		expect.RegisterSSLGenerationStep().Return(nil)
 		expect.RegisterValidationStep().Return(nil)
 		expect.RegisterComponentSetupSteps().Return(nil)
-		expect.RegisterDataSetupSteps(mock.Anything).Return(assert.AnError)
+		expect.RegisterDataSetupSteps(mock.Anything, mock.Anything).Return(assert.AnError)
 		starter.SetupExecutor = executorMock
 
 		// when
@@ -187,7 +187,7 @@ func TestStarter_StartSetup(t *testing.T) {
 		expect.RegisterSSLGenerationStep().Return(nil)
 		expect.RegisterValidationStep().Return(nil)
 		expect.RegisterComponentSetupSteps().Return(nil)
-		expect.RegisterDataSetupSteps(mock.Anything).Return(nil)
+		expect.RegisterDataSetupSteps(mock.Anything, mock.Anything).Return(nil)
 		expect.RegisterDoguInstallationSteps().Return(assert.AnError)
 		starter.SetupExecutor = executorMock
 
