@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package validation
 
@@ -23,6 +23,10 @@ func (_m *MockAdminValidator) EXPECT() *MockAdminValidator_Expecter {
 // ValidateAdmin provides a mock function with given fields: admin, dsType
 func (_m *MockAdminValidator) ValidateAdmin(admin context.User, dsType string) error {
 	ret := _m.Called(admin, dsType)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ValidateAdmin")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.User, string) error); ok {
@@ -63,13 +67,12 @@ func (_c *MockAdminValidator_ValidateAdmin_Call) RunAndReturn(run func(context.U
 	return _c
 }
 
-type mockConstructorTestingTNewMockAdminValidator interface {
+// NewMockAdminValidator creates a new instance of MockAdminValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockAdminValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockAdminValidator creates a new instance of MockAdminValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockAdminValidator(t mockConstructorTestingTNewMockAdminValidator) *MockAdminValidator {
+}) *MockAdminValidator {
 	mock := &MockAdminValidator{}
 	mock.Mock.Test(t)
 
