@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package validation
 
@@ -23,6 +23,10 @@ func (_m *MockDoguValidator) EXPECT() *MockDoguValidator_Expecter {
 // ValidateDogus provides a mock function with given fields: dogus
 func (_m *MockDoguValidator) ValidateDogus(dogus context.Dogus) error {
 	ret := _m.Called(dogus)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ValidateDogus")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Dogus) error); ok {
@@ -62,13 +66,12 @@ func (_c *MockDoguValidator_ValidateDogus_Call) RunAndReturn(run func(context.Do
 	return _c
 }
 
-type mockConstructorTestingTNewMockDoguValidator interface {
+// NewMockDoguValidator creates a new instance of MockDoguValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockDoguValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockDoguValidator creates a new instance of MockDoguValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockDoguValidator(t mockConstructorTestingTNewMockDoguValidator) *MockDoguValidator {
+}) *MockDoguValidator {
 	mock := &MockDoguValidator{}
 	mock.Mock.Test(t)
 
