@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package validation
 
@@ -23,6 +23,10 @@ func (_m *MockNamingValidator) EXPECT() *MockNamingValidator_Expecter {
 // ValidateNaming provides a mock function with given fields: naming
 func (_m *MockNamingValidator) ValidateNaming(naming context.Naming) error {
 	ret := _m.Called(naming)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ValidateNaming")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Naming) error); ok {
@@ -62,13 +66,12 @@ func (_c *MockNamingValidator_ValidateNaming_Call) RunAndReturn(run func(context
 	return _c
 }
 
-type mockConstructorTestingTNewMockNamingValidator interface {
+// NewMockNamingValidator creates a new instance of MockNamingValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockNamingValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockNamingValidator creates a new instance of MockNamingValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockNamingValidator(t mockConstructorTestingTNewMockNamingValidator) *MockNamingValidator {
+}) *MockNamingValidator {
 	mock := &MockNamingValidator{}
 	mock.Mock.Test(t)
 
