@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package validation
 
@@ -23,6 +23,10 @@ func (_m *MockRegistryConfigEncryptedValidator) EXPECT() *MockRegistryConfigEncr
 // ValidateRegistryConfigEncrypted provides a mock function with given fields: config
 func (_m *MockRegistryConfigEncryptedValidator) ValidateRegistryConfigEncrypted(config *context.SetupJsonConfiguration) error {
 	ret := _m.Called(config)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ValidateRegistryConfigEncrypted")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(*context.SetupJsonConfiguration) error); ok {
@@ -62,13 +66,12 @@ func (_c *MockRegistryConfigEncryptedValidator_ValidateRegistryConfigEncrypted_C
 	return _c
 }
 
-type mockConstructorTestingTNewMockRegistryConfigEncryptedValidator interface {
+// NewMockRegistryConfigEncryptedValidator creates a new instance of MockRegistryConfigEncryptedValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockRegistryConfigEncryptedValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockRegistryConfigEncryptedValidator creates a new instance of MockRegistryConfigEncryptedValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockRegistryConfigEncryptedValidator(t mockConstructorTestingTNewMockRegistryConfigEncryptedValidator) *MockRegistryConfigEncryptedValidator {
+}) *MockRegistryConfigEncryptedValidator {
 	mock := &MockRegistryConfigEncryptedValidator{}
 	mock.Mock.Test(t)
 
