@@ -1,4 +1,4 @@
-// Code generated by mockery v2.20.0. DO NOT EDIT.
+// Code generated by mockery v2.42.1. DO NOT EDIT.
 
 package validation
 
@@ -23,6 +23,10 @@ func (_m *MockUserBackendValidator) EXPECT() *MockUserBackendValidator_Expecter 
 // ValidateUserBackend provides a mock function with given fields: backend
 func (_m *MockUserBackendValidator) ValidateUserBackend(backend context.UserBackend) error {
 	ret := _m.Called(backend)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ValidateUserBackend")
+	}
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.UserBackend) error); ok {
@@ -62,13 +66,12 @@ func (_c *MockUserBackendValidator_ValidateUserBackend_Call) RunAndReturn(run fu
 	return _c
 }
 
-type mockConstructorTestingTNewMockUserBackendValidator interface {
+// NewMockUserBackendValidator creates a new instance of MockUserBackendValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockUserBackendValidator(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewMockUserBackendValidator creates a new instance of MockUserBackendValidator. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewMockUserBackendValidator(t mockConstructorTestingTNewMockUserBackendValidator) *MockUserBackendValidator {
+}) *MockUserBackendValidator {
 	mock := &MockUserBackendValidator{}
 	mock.Mock.Test(t)
 
